Remove Jilt integration & deprecate functions #8290
