+++ assets/js/admin/settings/index.js
@@ -1,5 +1,4 @@
 import { sendwpRemoteInstall, sendwpDisconnect } from './sendwp';
-import { jiltRemoteInstall, jiltDisconnect } from './jilt';
 
 /**
  * Settings screen JS
@@ -166,20 +165,6 @@ const EDD_Settings = {
 			$(this).html( edd_vars.wait + ' <span class="edd-loading dark"></span>' );
 			document.body.style.cursor = 'wait';
 			sendwpDisconnect();
-		});
-
-		$('#edd-jilt-connect').on('click', function(e) {
-			e.preventDefault();
-			$(this).html( edd_vars.wait + ' <span class="edd-loading"></span>' );
-			document.body.style.cursor = 'wait';
-			jiltRemoteInstall();
-		});
-
-		$('#edd-jilt-disconnect').on('click', function(e) {
-			e.preventDefault();
-			$(this).html( edd_vars.wait + ' <span class="edd-loading dark"></span>' );
-			document.body.style.cursor = 'wait';
-			jiltDisconnect();
 		});
 	}
 };
