+++ assets/js/edd-admin-settings.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function n(o){if(e[o])return e[o].exports;var i=e[o]={i:o,l:!1,exports:{}};return t[o].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=t,n.c=e,n.d=function(t,e,o){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:o})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var o=Object.create(null);if(n.r(o),Object.defineProperty(o,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)n.d(o,i,function(e){return t[e]}.bind(null,i));return o},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=61)}({0:function(t,e){t.exports=jQuery},22:function(t,e,n){"use strict";(function(t){n.d(e,"b",(function(){return r})),n.d(e,"a",(function(){return a}));var o=n(32),i=n.n(o),r=function(){t.post(ajaxurl,{action:"edd_sendwp_remote_install"},(function(t){t.success||!confirm(t.data.error)?d(t.data.register_url,t.data.client_name,t.data.client_secret,t.data.client_redirect,t.data.partner_id):location.reload()}))},a=function(){t.post(ajaxurl,{action:"edd_sendwp_disconnect"},(function(t){location.reload()}))},d=function(t,e,n,o,r){var a=document.createElement("form");a.setAttribute("method","POST"),a.setAttribute("action",t);for(var d=0,c=[["client_name",e],["client_secret",n],["client_redirect",o],["partner_id",r]];d<c.length;d++){var u=i()(c[d],2),l=u[0],s=u[1],f=document.createElement("input");f.setAttribute("type","hidden"),f.setAttribute("name",l),f.setAttribute("value",s),a.appendChild(f)}document.body.appendChild(a),a.submit()}}).call(this,n(0))},23:function(t,e,n){"use strict";(function(t){n.d(e,"b",(function(){return o})),n.d(e,"a",(function(){return r}));var o=function(){t.post(ajaxurl,{action:"edd_jilt_remote_install"},(function(t){t.success||!confirm(t.data.error)?i():location.reload()}))},i=function(){t.post(ajaxurl,{action:"edd_jilt_connect"},(function(t){t.success||!confirm(t.data.error)?""===t.data.connect_url||location.assign(t.data.connect_url):location.reload()}))},r=function(){t.post(ajaxurl,{action:"edd_jilt_disconnect"},(function(t){t.success||confirm(t.data.error),location.reload()}))}}).call(this,n(0))},32:function(t,e,n){var o=n(62),i=n(63),r=n(64);t.exports=function(t,e){return o(t)||i(t,e)||r()}},61:function(t,e,n){"use strict";n.r(e),function(t,e){var o=n(22),i=n(23),r={init:function(){this.general(),this.misc(),this.gateways(),this.emails()},general:function(){var n=t(".edd-color-picker");if(n.length&&n.wpColorPicker(),"undefined"==typeof wp||"1"!==edd_vars.new_media_ui){var o=t(".edd_settings_upload_button");o.length>0&&(window.formfield="",t(document.body).on("click",o,(function(n){n.preventDefault(),window.formfield=t(this).parent().prev(),window.tbframe_interval=setInterval((function(){e("#TB_iframeContent").contents().find(".savesend .button").val(edd_vars.use_this_file).end().find("#insert-gallery, .wp-post-thumbnail").hide()}),2e3),tb_show(edd_vars.add_new_download,"media-upload.php?TB_iframe=true")})),window.edd_send_to_editor=window.send_to_editor,window.send_to_editor=function(e){window.formfield?(imgurl=t("a","<div>"+e+"</div>").attr("href"),window.formfield.val(imgurl),window.clearInterval(window.tbframe_interval),tb_remove()):window.edd_send_to_editor(e),window.send_to_editor=window.edd_send_to_editor,window.formfield="",window.imagefield=!1})}else{var i;window.formfield="",t(document.body).on("click",".edd_settings_upload_button",(function(e){e.preventDefault();var n=t(this);window.formfield=t(this).parent().prev(),i||((i=wp.media.frames.file_frame=wp.media({title:n.data("uploader_title"),library:{type:"image"},button:{text:n.data("uploader_button_text")},multiple:!1})).on("menu:render:default",(function(t){t.unset("library-separator"),t.unset("gallery"),t.unset("featured-image"),t.unset("embed"),t.set({})})),i.on("select",(function(){i.state().get("selection").each((function(t,e){t=t.toJSON(),window.formfield.val(t.url)}))}))),i.open()})),window.formfield=""}},misc:function(){var e=t('select[name="edd_settings[download_method]"]'),n=e.parent().parent().next();"direct"===e.val()&&(n.css("opacity","0.4"),n.find("input").prop("checked",!1).prop("disabled",!0)),e.on("change",(function(){"direct"===t(this).val()?(n.css("opacity","0.4"),n.find("input").prop("checked",!1).prop("disabled",!0)):(n.find("input").prop("disabled",!1),n.css("opacity","1"))}))},gateways:function(){t('#edd-payment-gateways input[type="checkbox"]').on("change",(function(){var e=t(this).data("gateway-key"),n=t("#edd_settings\\[default_gateway\\]"),o=n.find('option[value="'+e+'"]');o.prop("disabled",(function(t,e){return!e})),o.prop("selected")&&o.prop("selected",!1),n.trigger("chosen:updated")}))},emails:function(){t("#edd-sendwp-connect").on("click",(function(e){e.preventDefault(),t(this).html(edd_vars.wait+' <span class="edd-loading"></span>'),document.body.style.cursor="wait",Object(o.b)()})),t("#edd-sendwp-disconnect").on("click",(function(e){e.preventDefault(),t(this).html(edd_vars.wait+' <span class="edd-loading dark"></span>'),document.body.style.cursor="wait",Object(o.a)()})),t("#edd-jilt-connect").on("click",(function(e){e.preventDefault(),t(this).html(edd_vars.wait+' <span class="edd-loading"></span>'),document.body.style.cursor="wait",Object(i.b)()})),t("#edd-jilt-disconnect").on("click",(function(e){e.preventDefault(),t(this).html(edd_vars.wait+' <span class="edd-loading dark"></span>'),document.body.style.cursor="wait",Object(i.a)()}))}};e(document).ready((function(t){r.init()}))}.call(this,n(0),n(0))},62:function(t,e){t.exports=function(t){if(Array.isArray(t))return t}},63:function(t,e){t.exports=function(t,e){if(Symbol.iterator in Object(t)||"[object Arguments]"===Object.prototype.toString.call(t)){var n=[],o=!0,i=!1,r=void 0;try{for(var a,d=t[Symbol.iterator]();!(o=(a=d.next()).done)&&(n.push(a.value),!e||n.length!==e);o=!0);}catch(t){i=!0,r=t}finally{try{o||null==d.return||d.return()}finally{if(i)throw r}}return n}}},64:function(t,e){t.exports=function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}}});
+!function(e){var t={};function n(i){if(t[i])return t[i].exports;var r=t[i]={i:i,l:!1,exports:{}};return e[i].call(r.exports,r,r.exports,n),r.l=!0,r.exports}n.m=e,n.c=t,n.d=function(e,t,i){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:i})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var i=Object.create(null);if(n.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)n.d(i,r,function(t){return e[t]}.bind(null,r));return i},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=61)}({0:function(e,t){e.exports=jQuery},22:function(e,t,n){"use strict";(function(e){n.d(t,"b",(function(){return o})),n.d(t,"a",(function(){return d}));var i=n(32),r=n.n(i),o=function(){e.post(ajaxurl,{action:"edd_sendwp_remote_install"},(function(e){e.success||!confirm(e.data.error)?a(e.data.register_url,e.data.client_name,e.data.client_secret,e.data.client_redirect,e.data.partner_id):location.reload()}))},d=function(){e.post(ajaxurl,{action:"edd_sendwp_disconnect"},(function(e){location.reload()}))},a=function(e,t,n,i,o){var d=document.createElement("form");d.setAttribute("method","POST"),d.setAttribute("action",e);for(var a=0,c=[["client_name",t],["client_secret",n],["client_redirect",i],["partner_id",o]];a<c.length;a++){var u=r()(c[a],2),l=u[0],s=u[1],f=document.createElement("input");f.setAttribute("type","hidden"),f.setAttribute("name",l),f.setAttribute("value",s),d.appendChild(f)}document.body.appendChild(d),d.submit()}}).call(this,n(0))},32:function(e,t,n){var i=n(62),r=n(63),o=n(64);e.exports=function(e,t){return i(e)||r(e,t)||o()}},61:function(e,t,n){"use strict";n.r(t),function(e,t){var i=n(22),r={init:function(){this.general(),this.misc(),this.gateways(),this.emails()},general:function(){var n=e(".edd-color-picker");if(n.length&&n.wpColorPicker(),"undefined"==typeof wp||"1"!==edd_vars.new_media_ui){var i=e(".edd_settings_upload_button");i.length>0&&(window.formfield="",e(document.body).on("click",i,(function(n){n.preventDefault(),window.formfield=e(this).parent().prev(),window.tbframe_interval=setInterval((function(){t("#TB_iframeContent").contents().find(".savesend .button").val(edd_vars.use_this_file).end().find("#insert-gallery, .wp-post-thumbnail").hide()}),2e3),tb_show(edd_vars.add_new_download,"media-upload.php?TB_iframe=true")})),window.edd_send_to_editor=window.send_to_editor,window.send_to_editor=function(t){window.formfield?(imgurl=e("a","<div>"+t+"</div>").attr("href"),window.formfield.val(imgurl),window.clearInterval(window.tbframe_interval),tb_remove()):window.edd_send_to_editor(t),window.send_to_editor=window.edd_send_to_editor,window.formfield="",window.imagefield=!1})}else{var r;window.formfield="",e(document.body).on("click",".edd_settings_upload_button",(function(t){t.preventDefault();var n=e(this);window.formfield=e(this).parent().prev(),r||((r=wp.media.frames.file_frame=wp.media({title:n.data("uploader_title"),library:{type:"image"},button:{text:n.data("uploader_button_text")},multiple:!1})).on("menu:render:default",(function(e){e.unset("library-separator"),e.unset("gallery"),e.unset("featured-image"),e.unset("embed"),e.set({})})),r.on("select",(function(){r.state().get("selection").each((function(e,t){e=e.toJSON(),window.formfield.val(e.url)}))}))),r.open()})),window.formfield=""}},misc:function(){var t=e('select[name="edd_settings[download_method]"]'),n=t.parent().parent().next();"direct"===t.val()&&(n.css("opacity","0.4"),n.find("input").prop("checked",!1).prop("disabled",!0)),t.on("change",(function(){"direct"===e(this).val()?(n.css("opacity","0.4"),n.find("input").prop("checked",!1).prop("disabled",!0)):(n.find("input").prop("disabled",!1),n.css("opacity","1"))}))},gateways:function(){e('#edd-payment-gateways input[type="checkbox"]').on("change",(function(){var t=e(this).data("gateway-key"),n=e("#edd_settings\\[default_gateway\\]"),i=n.find('option[value="'+t+'"]');i.prop("disabled",(function(e,t){return!t})),i.prop("selected")&&i.prop("selected",!1),n.trigger("chosen:updated")}))},emails:function(){e("#edd-sendwp-connect").on("click",(function(t){t.preventDefault(),e(this).html(edd_vars.wait+' <span class="edd-loading"></span>'),document.body.style.cursor="wait",Object(i.b)()})),e("#edd-sendwp-disconnect").on("click",(function(t){t.preventDefault(),e(this).html(edd_vars.wait+' <span class="edd-loading dark"></span>'),document.body.style.cursor="wait",Object(i.a)()}))}};t(document).ready((function(e){r.init()}))}.call(this,n(0),n(0))},62:function(e,t){e.exports=function(e){if(Array.isArray(e))return e}},63:function(e,t){e.exports=function(e,t){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e)){var n=[],i=!0,r=!1,o=void 0;try{for(var d,a=e[Symbol.iterator]();!(i=(d=a.next()).done)&&(n.push(d.value),!t||n.length!==t);i=!0);}catch(e){r=!0,o=e}finally{try{i||null==a.return||a.return()}finally{if(r)throw o}}return n}}},64:function(e,t){e.exports=function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}}});
