+++ includes/admin/settings/register-settings.php
@@ -673,17 +673,6 @@ function edd_get_registered_settings() {
 						'desc'    => '',
 						'type'    => 'hook',
 					),
-					'advanced_emails_header' => array(
-						'id'   => 'advanced_emails_header',
-						'name' => '<strong>' . __( 'Advanced emails', 'easy-digital-downloads' ) . '</strong>',
-						'type' => 'header',
-					),
-					'jilt'                   => array(
-						'id'   => 'jilt',
-						'name' => __( 'Enhanced emails via Jilt', 'easy-digital-downloads' ),
-						'desc' => '<p>' . __( 'Create beautiful transactional, automated, and marketing emails using a drag-and-drop editor with <a href="https://jilt.com/?utm_source=edd-core&utm_medium=referral&utm_campaign=edd-enhanced-emails" target="_blank" rel="noopener noreferrer">Jilt</a>.', 'easy-digital-downloads' ) . '</p><p>' . __( 'Learn more about free and paid plans in the <a href="https://docs.easydigitaldownloads.com/article/2199-jilt-overview" target="_blank" rel="noopener noreferrer">documentation</a>.', 'easy-digital-downloads' ) . '</p>',
-						'type' => 'jilt',
-					),
 				),
 				'purchase_receipts' => array(
 					'purchase_receipt_email_settings' => array(
@@ -2665,68 +2654,6 @@ function edd_sendwp_callback($args) {
 function edd_tax_rate_callback( $args ) {
 	echo '<input type="hidden" id="edd_settings[' . edd_sanitize_key( $args['id'] ) . ']" name="edd_settings[' . esc_attr( $args['id'] ) . ']" value="" />';
 	echo wp_kses_post( $args['desc'] );
-}
-
-/**
- * Jilt Callback
- *
- * Renders Jilt Settings
- *
- * @since n.n.n
- * @param array $args arguments passed by the setting.
- * @return void
- */
-function edd_jilt_callback( $args ) {
-
-	$activated   = is_callable( 'edd_jilt' );
-	$connected   = $activated && edd_jilt()->get_integration()->is_jilt_connected();
-	$connect_url = $activated ? edd_jilt()->get_connect_url() : '';
-	$account_url = $connected ? edd_jilt()->get_integration()->get_jilt_app_url() : '';
-
-	echo wp_kses_post( $args['desc'] );
-
-	if ( $activated ) :
-		?>
-
-		<?php if ( $connected ) : ?>
-
-			<p>
-				<button id="edd-jilt-disconnect" class="button"><?php esc_html_e( 'Disconnect Jilt', 'easy-digital-downloads' ); ?></button>
-			</p>
-
-			<p>
-				<?php
-				wp_kses_post(
-					sprintf(
-						/* Translators: %1$s - <a> tag, %2$s - </a> tag */
-						__( '%1$sClick here%2$s to visit your Jilt dashboard', 'easy-digital-downloads' ),
-						'<a href="' . esc_url( $account_url ) . '" target="_blank">',
-						'</a>'
-					)
-				);
-				?>
-			</p>
-
-		<?php else : ?>
-
-			<p>
-				<a id="edd-jilt-connect" class="button button-primary" href="<?php echo esc_url( $connect_url ); ?>">
-					<?php esc_html_e( 'Connect to Jilt', 'easy-digital-downloads' ); ?>
-				</a>
-			</p>
-
-		<?php endif; ?>
-
-	<?php elseif( current_user_can( 'install_plugins' ) ) : ?>
-
-		<p>
-			<button id="edd-jilt-connect" class="button button-primary">
-				<?php esc_html_e( 'Install Jilt', 'easy-digital-downloads' ); ?>
-			</button>
-		</p>
-
-	<?php
-	endif;
 }
 
 /**
