+++ includes/deprecated-functions.php
@@ -1132,3 +1132,204 @@ function edd_record_sale_in_log( $download_id = 0, $payment_id, $price_id = fals
 function edd_agree_to_terms_js() {
 	_edd_deprecated_function( __FUNCTION__, '3.0' );
 }
+
+/**
+ * Handle installation and activation for Jilt via AJAX
+ *
+ * @deprecated 3.0
+ *
+ * @since n.n.n
+ */
+function edd_jilt_remote_install_handler() {
+
+	_edd_deprecated_function( __FUNCTION__, '3.0' );
+
+	if ( ! current_user_can( 'manage_shop_settings' ) || ! current_user_can( 'install_plugins' ) ) {
+		wp_send_json_error(
+				array(
+						'error' => __( 'You do not have permission to do this.', 'easy-digital-downloads' ),
+				)
+		);
+	}
+
+	include_once ABSPATH . 'wp-admin/includes/plugin-install.php';
+	include_once ABSPATH . 'wp-admin/includes/file.php';
+	include_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
+
+	$plugins = get_plugins();
+
+	if ( ! array_key_exists( 'jilt-for-edd/jilt-for-edd.php', $plugins ) ) {
+		/*
+		* Use the WordPress Plugins API to get the plugin download link.
+		*/
+		$api = plugins_api(
+				'plugin_information',
+				array(
+						'slug' => 'jilt-for-edd',
+				)
+		);
+
+		if ( is_wp_error( $api ) ) {
+			wp_send_json_error(
+					array(
+							'error' => $api->get_error_message(),
+							'debug' => $api,
+					)
+			);
+		}
+
+		/*
+		* Use the AJAX Upgrader skin to quietly install the plugin.
+		*/
+		$upgrader = new Plugin_Upgrader( new WP_Ajax_Upgrader_Skin() );
+		$install  = $upgrader->install( $api->download_link );
+		if ( is_wp_error( $install ) ) {
+			wp_send_json_error(
+					array(
+							'error' => $install->get_error_message(),
+							'debug' => $api,
+					)
+			);
+		}
+
+		activate_plugin( $upgrader->plugin_info() );
+
+	} else {
+
+		activate_plugin( 'jilt-for-edd/jilt-for-edd.php' );
+	}
+
+	/*
+	* Final check to see if Jilt is available.
+	*/
+	if ( ! class_exists( 'EDD_Jilt_Loader' ) ) {
+		wp_send_json_error(
+				array(
+						'error' => __( 'Something went wrong. Jilt was not installed correctly.', 'easy-digital-downloads' ),
+				)
+		);
+	}
+
+	wp_send_json_success();
+}
+
+/**
+ * Handle connection for Jilt via AJAX
+ *
+ * @deprecated 3.0
+ *
+ * @since n.n.n
+ */
+function edd_jilt_connect_handler() {
+
+	_edd_deprecated_function( __FUNCTION__, '3.0' );
+
+	if ( ! current_user_can( 'manage_shop_settings' ) ) {
+		wp_send_json_error(
+				array(
+						'error' => __( 'You do not have permission to do this.', 'easy-digital-downloads' ),
+				)
+		);
+	}
+
+	if ( ! is_callable( 'edd_jilt' ) ) {
+		wp_send_json_error(
+				array(
+						'error' => __( 'Something went wrong. Jilt was not installed correctly.', 'easy-digital-downloads' ),
+				)
+		);
+	}
+
+	wp_send_json_success( array( 'connect_url' => edd_jilt()->get_connect_url() ) );
+}
+
+/**
+ * Handle disconnection and deactivation for Jilt via AJAX
+ *
+ * @deprecated 3.0
+ *
+ * @since n.n.n
+ */
+function edd_jilt_disconnect_handler() {
+
+	_edd_deprecated_function( __FUNCTION__, '3.0' );
+
+	if ( ! current_user_can( 'manage_shop_settings' ) ) {
+		wp_send_json_error(
+				array(
+						'error' => __( 'You do not have permission to do this.', 'easy-digital-downloads' ),
+				)
+		);
+	}
+
+	if ( is_callable( 'edd_jilt' ) ) {
+
+		edd_jilt()->get_integration()->unlink_shop();
+		edd_jilt()->get_integration()->revoke_authorization();
+		edd_jilt()->get_integration()->clear_connection_data();
+	}
+
+	deactivate_plugins( 'jilt-for-edd/jilt-for-edd.php' );
+
+	wp_send_json_success();
+}
+
+/**
+ * Maybe adds a notice to abandoned payments if Jilt isn't installed.
+ *
+ * @deprecated 3.0
+ *
+ * @since n.n.n
+ *
+ * @param int $payment_id The ID of the abandoned payment, for which a jilt notice is being thrown.
+ */
+function maybe_add_jilt_notice_to_abandoned_payment( $payment_id ) {
+
+	_edd_deprecated_function( __FUNCTION__, '3.0' );
+
+	if ( ! is_callable( 'edd_jilt' )
+			&& ! is_plugin_active( 'recapture-for-edd/recapture.php' )
+			&& 'abandoned' === edd_get_payment_status( $payment_id )
+			&& ! get_user_meta( get_current_user_id(), '_edd_try_jilt_dismissed', true )
+	) {
+		?>
+		<div class="notice notice-warning jilt-notice">
+			<p>
+				<?php
+				echo wp_kses_post(
+						sprintf(
+						/* Translators: %1$s - <strong> tag, %2$s - </strong> tag, %3$s - <a> tag, %4$s - </a> tag */
+								__( '%1$sRecover abandoned purchases like this one.%2$s %3$sTry Jilt for free%4$s.', 'easy-digital-downloads' ),
+								'<strong>',
+								'</strong>',
+								'<a href="https://easydigitaldownloads.com/downloads/jilt" target="_blank">',
+								'</a>'
+						)
+				);
+				?>
+			</p>
+			<?php
+			echo wp_kses_post(
+					sprintf(
+					/* Translators: %1$s - Opening anchor tag, %2$s - The url to dismiss the ajax notice, %3$s - Complete the opening of the anchor tag, %4$s - Open span tag, %4$s - Close span tag */
+							__( '%1$s %2$s %3$s %4$s Dismiss this notice. %5$s', 'easy-digital-downloads' ),
+							'<a href="',
+							esc_url(
+									add_query_arg(
+											array(
+													'edd_action' => 'dismiss_notices',
+													'edd_notice' => 'try_jilt',
+											)
+									)
+							),
+							'" type="button" class="notice-dismiss">',
+							'<span class="screen-reader-text">',
+							'</span>
+					</a>'
+					)
+			);
+			?>
+		</div>
+		<?php
+	}
+}
